Clean the code up and hook up garbage collection
