+++ AirCasting/Utils/Logging/FileLogger/FileLogger.swift
@@ -3,14 +3,20 @@
 
 import Foundation
 
+/// Use d by the `FileLogger` for opening log files
 protocol FileLoggerStore {
     func openOrCreateLogFile() throws -> FileLoggerFileHandle
 }
 
+/// Returned by the `FileLoggerStore`. Used for appending previously opened file
+/// - Note:
+/// The system is designed so that file closes after releasing this handle from the memory.
 protocol FileLoggerFileHandle {
+    /// Appends the file with a given string
     func appendFile(with: String) throws
 }
 
+/// A logger that outputs logs into a file
 class FileLogger: Logger {
     private let store: FileLoggerStore
     private let fileHandle: FileLoggerFileHandle
